% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_lock_storr.R
\name{multi_lock}
\alias{multi_lock}
\title{Multiple locks}
\usage{
multi_lock(path, allowed_number_of_jobs, id, fresh = F)
}
\arguments{
\item{path}{A local filesystem path where lock will be created}

\item{allowed_number_of_jobs}{Maximum number of jobs (tasks or processes) which is possible to run concurrently. The default is 1 (unless any process initiated more than that)}

\item{id}{An identifier for uniquely pointing the lock (optional parameter).}

\item{fresh}{Default is \code{FALSE}. If \code{TRUE} the directory (as specified by path) will be cleaned and recreated so that earlier locks will not interfere.}
}
\value{
An access handle (environment) with lock, unlock and destroy method. All are empty argument functions.
\code{lock} will try to lock the current process.
\code{unlock} will unlock the current process.
\code{destroy} will delete the lock folder. (Potentially this should be called by only one process at the end of operations)
}
\description{
Multiple locks
}
\examples{

}
